backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
asyncpg==0.30.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-keycloak==3.9.1

# Background tasks
redis==5.0.1
rq==1.15.1

# Excel processing (for Sprint 1)
openpyxl==3.1.2
pandas==2.1.4

# CLI tools
click==8.1.7

# LangChain & AI processing (Sprint 3)
langchain==0.2.16
langchain-community==0.2.16
langchain-postgres==0.0.15
langchain-huggingface==0.0.3
psycopg-binary==3.2.1
sentence-transformers==2.7.0
transformers==4.40.0
huggingface-hub==0.24.6
pypdf==3.17.4
python-docx==1.1.0
pgvector==0.2.5
ollama==0.1.7
langdetect==1.0.9

# Utils
httpx==0.25.2
structlog==24.1.0

# PDF generation
reportlab==4.0.9
jinja2==3.1.2
weasyprint==66.0
pydyf==0.11.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==23.12.1
ruff==0.1.11
mypy==1.8.0
pre-commit==3.6.0